chart working.
looking up full type from pokemon directly before determining their
effectiveness. should be safe since typedex will have to exist before
function is called, but slapping the type and the pokemon together as
part of building the table is really messy and adds a lot of complexity
in the middle of something that is automatically somewhat complex.