Switch to svelte (#2)
Elm was working but getting really messy/complex due to the number. Switched to svelte. It was pretty much that or try rewriting it without the elm plugin I was using for the autocompletes which seemed to add more boilerplate than it did functionality.

Added browser side caching of pokeapi.